relocate context menu provider
